fix: Fail on invalid factory keys.
Fixes #1260
